+++ bytebank/src/main/kotlin/br/com/alura/bytebank/main.kt
@@ -1,21 +1,5 @@
-import br.com.alura.bytebank.modelo.Endereco
-
 fun main() {
-    var enderecoNulo: Endereco? = Endereco(logradouro = "Rua Vergueiro", complemento = "prédio")
-    val logradouroNovo: String? = enderecoNulo?.logradouro
-
-    enderecoNulo?.let { endereco ->
-        println(endereco.logradouro.length)
-        val tamanhoComplemento: Int =
-            endereco.complemento?.length ?: throw IllegalStateException("Complemento não pode ser vazio")
-        println(tamanhoComplemento)
-    }
-
-    teste(1)
-    teste("")
+    println(soma(1, 5))
 }
 
-fun teste(valor: Any){
-    val numero: Int? = valor as? Int
-    println(numero)
-}
+fun soma(a: Int, b: Int) : Int = a + b
